Separate repositories for different features by packages
