--- app/src/main/java/com/github/fo2rist/mclaren/repository/McLarenFeedRepository.java
@@ -1,4 +1,0 @@
-package com.github.fo2rist.mclaren.repository;
-
-public interface McLarenFeedRepository extends FeedRepository {
-}
