+++ app/src/main/java/com/github/fo2rist/mclaren/repository/feed/FeedRepository.kt
@@ -1,7 +1,7 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.feed
 
 import com.github.fo2rist.mclaren.models.FeedItem
-import com.github.fo2rist.mclaren.repository.FeedRepositoryEventBus.LoadingEvent
+import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus.LoadingEvent
 import com.github.fo2rist.mclaren.repository.converters.FeedConverter
 import com.github.fo2rist.mclaren.utils.toDescendingList
 import com.github.fo2rist.mclaren.web.DEFAULT_PAGE
