+++ app/src/main/java/com/github/fo2rist/mclaren/repository/feed/McLarenFeedRepository.java
@@ -1,0 +1,6 @@
+package com.github.fo2rist.mclaren.repository.feed;
+
+import com.github.fo2rist.mclaren.repository.feed.FeedRepository;
+
+public interface McLarenFeedRepository extends FeedRepository {
+}
