+++ app/src/main/java/com/github/fo2rist/mclaren/repository/feed/TwitterRepositoryImpl.kt
@@ -1,11 +1,12 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.feed
 
 import com.github.fo2rist.mclaren.models.FeedItem
 import com.github.fo2rist.mclaren.models.ImageUrl
 import com.github.fo2rist.mclaren.models.Size
-import com.github.fo2rist.mclaren.repository.FeedRepositoryEventBus.LoadingEvent
+import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus.LoadingEvent
+import com.github.fo2rist.mclaren.repository.feed.FeedRepository
+import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus
 import com.github.fo2rist.mclaren.utils.toDescendingList
-import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import twitter4j.Paging
 import javax.inject.Inject
