+++ app/src/main/java/com/github/fo2rist/mclaren/repository/remoteconfig/RaceCalendarRepository.kt
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.remoteconfig
 
 import com.github.fo2rist.mclaren.ui.models.RaceCalendar
