+++ app/src/main/java/com/github/fo2rist/mclaren/repository/transmission/TransmissionConverter.kt
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.transmission
 
 import com.github.fo2rist.mclaren.models.TransmissionInfo
 import com.github.fo2rist.mclaren.models.TransmissionItem
