+++ app/src/main/java/com/github/fo2rist/mclaren/repository/transmission/TransmissionRepositoryImpl.kt
@@ -1,7 +1,7 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.transmission
 
 import com.github.fo2rist.mclaren.models.TransmissionInfo
-import com.github.fo2rist.mclaren.repository.TransmissionRepositoryEventBus.LoadingEvent
+import com.github.fo2rist.mclaren.repository.transmission.TransmissionRepositoryEventBus.LoadingEvent
 import com.github.fo2rist.mclaren.web.SafeJsonParser
 import com.github.fo2rist.mclaren.web.TransmissionWebService
 import com.github.fo2rist.mclaren.web.models.Transmission
