+++ app/src/test/java/com/github/fo2rist/mclaren/repository/feed/McLarenFeedRepositoryImplTest.kt
@@ -1,6 +1,8 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.feed
 
-import com.github.fo2rist.mclaren.repository.FeedRepositoryEventBus.LoadingEvent
+import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus
+import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus.LoadingEvent
+import com.github.fo2rist.mclaren.repository.feed.McLarenFeedRepositoryImpl
 import com.github.fo2rist.mclaren.testdata.McLarenFeedResponse.REAL_FEED_RESPONSE
 import com.github.fo2rist.mclaren.web.FeedHistoryPredictor
 import com.github.fo2rist.mclaren.web.McLarenFeedWebService
