+++ app/src/test/java/com/github/fo2rist/mclaren/repository/feed/StoryStreamRepositoryImplTest.kt
@@ -1,6 +1,6 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.feed
 
-import com.github.fo2rist.mclaren.repository.FeedRepositoryEventBus.LoadingEvent
+import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus.LoadingEvent
 import com.github.fo2rist.mclaren.testdata.StoryStreamResponse.REAL_FEED_RESPONSE
 import com.github.fo2rist.mclaren.web.StoryStreamWebService
 import com.github.fo2rist.mclaren.web.utils.BadResponse
