+++ app/src/test/java/com/github/fo2rist/mclaren/repository/remoteconfig/DriversRepositoryImplTest.kt
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.remoteconfig
 
 import com.github.fo2rist.mclaren.testdata.DUMMY_URL
 import com.github.fo2rist.mclaren.testdata.FakeRemoteConfigService
