+++ app/src/test/java/com/github/fo2rist/mclaren/repository/transmission/TransmissionConverterTest.kt
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.repository
+package com.github.fo2rist.mclaren.repository.transmission
 
 import com.github.fo2rist.mclaren.models.TransmissionItem
 import com.github.fo2rist.mclaren.testdata.REAL_TRANSMISSION_RESPONSE
