+++ app/src/test/java/com/github/fo2rist/mclaren/ui/MainPresenterTest.java
@@ -3,8 +3,8 @@ package com.github.fo2rist.mclaren.ui;
 import com.github.fo2rist.mclaren.analytics.Events;
 import com.github.fo2rist.mclaren.analytics.EventsLogger;
 import com.github.fo2rist.mclaren.mvp.MainScreenContract;
-import com.github.fo2rist.mclaren.repository.RaceCalendarRepository;
 import com.github.fo2rist.mclaren.ui.models.CalendarEvent;
+import com.github.fo2rist.mclaren.repository.remoteconfig.RaceCalendarRepository;
 import com.github.fo2rist.mclaren.ui.models.RaceCalendar;
 import org.joda.time.DateTime;
 import org.junit.Before;
