+++ app/src/test/java/com/github/fo2rist/mclaren/ui/feedscreen/FeedPresenterTest.java
@@ -4,9 +4,9 @@ import com.github.fo2rist.mclaren.analytics.Events;
 import com.github.fo2rist.mclaren.analytics.EventsLogger;
 import com.github.fo2rist.mclaren.models.FeedItem;
 import com.github.fo2rist.mclaren.mvp.FeedContract;
-import com.github.fo2rist.mclaren.repository.FeedRepository;
-import com.github.fo2rist.mclaren.repository.FeedRepositoryEventBus;
-import com.github.fo2rist.mclaren.repository.FeedRepositoryEventBus.LoadingEvent;
+import com.github.fo2rist.mclaren.repository.feed.FeedRepository;
+import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus;
+import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus.LoadingEvent;
 import com.github.fo2rist.mclaren.testdata.FeedItems;
 import java.util.ArrayList;
 import java.util.List;
